Adding Initial Forces, Baseline Corrections for Drift
I've added a large code section which takes over if there is a section of the sine waveform that can't be fit, where a linear regression is fit and subtracted. This is meant to correct for any drifting in the system. I've also added a sheet in the output excel which contains the average force for the first 200 values of the force. Finally, there were some quality of life changes, the open output folder command at the end is now cross-platform (mac/linux/windows) and the enter key also works to submit the files.