Fixing SineStrands Math & Adding Calibration
I added an option in the GUI to add a calibration
file, where the delay between the force conduction.

I also adjusted the math for the viscous and
elastic moduli, where the f_in for the muscle has
the delay from the calibration file (fin-lin
PhaseShifts) subtracted from it. This new phase shift
is then used with amplitudes to calculate moduli.

Finally, I changed the signs for the amplitudes, so
that it uses the negative amplitudes (where work is
produced).